// client/src/types/nice.ts
export interface VerificationData {
  responseno: string;
  birthdate: string;
  gender: string;
  di: string;
  mobileco: string;
  ci: string;
  receivedata: string;
  mobileno: string;
  requestno: string;
  nationalinfo: string;
  authtype: string;
  sitecode: string;
  utf8_name: string;
  enctime: string;
  name: string;
  resultcode: string;
}